added project files for chating ( RabbitMQ messaging)
see the installation guide for  how to setup rabbitmq server and run the
files. Use the code in the script tags for our project which should let
two users communicate. Also, in the project i am passing the values of
users as url parameters. change that as per need of our project